mini-projet/health-service/requirements.txt
Flask==3.0.3
requests==2.32.3
PyJWT==2.9.0
flask-swagger-ui==4.11.1
